add typescript to api
