+++ .github/workflows/api.yaml
@@ -53,6 +53,29 @@ jobs:
       - name: Run linter
         run: yarn lint
 
+  api-build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v2
+
+      - name: Set up Node
+        uses: actions/setup-node@v2
+        with:
+          node-version: 12
+
+      - name: Cache dependencies
+        uses: actions/cache@v2
+        with:
+          path: api/node_modules
+          key: ${{ runner.os }}-cache
+
+      - name: Install dependencies
+        run: yarn
+
+      - name: Check types
+        run: yarn typecheck
+
   api-test:
     runs-on: ubuntu-latest
     env:
