+++ .gitignore
@@ -105,3 +105,4 @@ dist
 
 build
 .vercel
+dist
