+++ api/package.json
@@ -3,16 +3,21 @@
   "version": "0.1.0",
   "license": "MIT",
   "scripts": {
-    "start": "if-env NODE_ENV=production ?? npm run start:prod || npm run start:dev",
-    "start:prod": "node -r dotenv/config ./bin/www",
-    "start:dev": "nodemon -r dotenv/config ./bin/www",
-    "lint": "eslint src",
-    "lint:fix": "eslint --fix src",
-    "format": "prettier --write \"./**/*.{js,jsx,json,md}\"",
-    "format:check": "prettier --check \"./**/*.{js,jsx,json,md}\"",
-    "test": "jest --setupFiles dotenv/config"
+    "start": "tsc && if-env NODE_ENV=production ?? npm run start:prod || npm run start:dev",
+    "start:prod": "node -r dotenv/config ./bin/www.js",
+    "start:dev": "nodemon -r dotenv/config ./bin/www.js",
+    "lint": "eslint src --ext ts",
+    "lint:fix": "eslint --fix src --ext ts",
+    "format": "prettier --write \"./**/*.{ts,json,md}\"",
+    "format:check": "prettier --check \"./**/*.{ts,json,md}\"",
+    "test": "jest --setupFiles dotenv/config",
+    "build": "tsc"
   },
   "dependencies": {
+    "@types/cors": "^2.8.9",
+    "@types/express": "^4.17.11",
+    "@types/http-errors": "^1.8.0",
+    "@types/morgan": "^1.9.2",
     "cors": "^2.8.5",
     "debug": "~4.3.1",
     "dotenv": "^8.2.0",
@@ -27,11 +32,13 @@
   },
   "devDependencies": {
     "@hack4impact-uiuc/eslint-plugin": "^2.0.9",
+    "@types/node": "^14.14.26",
     "eslint": "^7.18.0",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7",
     "prettier": "^2.2.1",
-    "supertest": "^6.1.1"
+    "supertest": "^6.1.1",
+    "typescript": "^4.1.5"
   },
   "jest": {
     "testEnvironment": "node",
