+++ api/src/api/index.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import users from './users';
+
+const router = express.Router();
+
+router.use('/users', users);
+
+export default router;
