--- api/src/middleware/errorWrap.js
@@ -1,8 +1,0 @@
-/*
- * Middleware to allow us to safely handle errors in async/await code
- * without wrapping each route in try...catch blocks
- */
-const errorWrap = (fn) => (req, res, next) =>
-  Promise.resolve(fn(req, res, next)).catch(next);
-
-module.exports = errorWrap;
