--- api/src/middleware/index.js
@@ -1,7 +1,0 @@
-const errorHandler = require('./errorHandler');
-const errorWrap = require('./errorWrap');
-
-module.exports = {
-  errorHandler,
-  errorWrap,
-};
