+++ api/src/middleware/index.ts
@@ -1,0 +1,4 @@
+import errorHandler from './errorHandler';
+import errorWrap from './errorWrap';
+
+export { errorHandler, errorWrap };
