--- api/src/models/user.js
@@ -1,7 +1,0 @@
-const mongoose = require('mongoose');
-
-const User = new mongoose.Schema({
-  name: { type: String },
-});
-
-module.exports = mongoose.model('User', User);
